Rename parser file
