+++ auctionator_parser.py
